requirements.txt
blinker==1.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
Flask==3.1.0
fonttools==4.55.4
fsspec==2024.12.0
gunicorn==23.0.0
huggingface-hub==0.27.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.2.0
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
safetensors==0.5.2
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
timm==1.0.14
tokenizers==0.21.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.9.0
tzdata==2025.1
ultralytics==8.3.65
ultralytics-thop==2.0.14
urllib3==1.26.13
Werkzeug==3.1.3
